Rename globals.py to constants.py
The module only holds an immutable shared constant (DECIMAL_ZERO), so
"constants" describes its purpose better than "globals", which implies
mutable global state. Updates all imports, .coveragerc, and CLAUDE.md
references, and replaces the open TODO with a clarifying comment.
