+++ .coveragerc
@@ -6,4 +6,4 @@ omit =
     .venv/*
     ./source/__init__.py
     ./source/color_theme.py
-    ./source/globals.py
+    ./source/constants.py
