+++ source/Invoice.py
@@ -1,4 +1,4 @@
-from source.globals import DECIMAL_ZERO
+from source.constants import DECIMAL_ZERO
 
 
 # Invoice class to hold all attributes of the invoice. This represents a single invoice generated by Fishbowl
