+++ source/InvoiceProcessor.py
@@ -9,7 +9,7 @@ from source.processor_utilities import (
 )
 from source.InvoiceAppFileIO import InvoiceAppFileIO
 from source.Invoice import Invoice
-from source.globals import DECIMAL_ZERO
+from source.constants import DECIMAL_ZERO
 
 
 # InvoiceProcessor class to handle all logic for text processing on invoices
