--- source/globals.py
@@ -1,5 +1,0 @@
-from decimal import Decimal
-
-# TODO: Do I even need this file? Is there a better way to do globals?
-
-DECIMAL_ZERO = Decimal("0.00")
