+++ tests/InvoiceProcessor_tests.py
@@ -5,7 +5,7 @@ from decimal import Decimal
 from source.InvoiceProcessor import InvoiceProcessor
 from source.InvoiceAppFileIO import InvoiceAppFileIO
 from source.Invoice import Invoice
-from source.globals import DECIMAL_ZERO
+from source.constants import DECIMAL_ZERO
 
 
 ###############################################################################
